backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
asyncpg==0.30.0
psycopg[binary]==3.2.9
celery==5.5.3
redis==6.4.0
pydantic-settings==2.10.1
pwdlib[argon2]==0.2.1
PyJWT==2.10.1
email-validator==2.2.0
openai>=1.40.0,<2.0.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
langchain>=0.2.0,<0.4.0
llama-index-core>=0.10.0,<0.13.0
python-dotenv==1.1.1
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
